requirements.txt
PyAudio
av
scipy
websocket-client
loguru